Move assets to own directory.
