+++ example-wasm/src/assets/logo.png
